Add solution for problem `C - Three Bags` Link: https://codeforces.com/contest/1467/problem/C Tags: constructive algorithms, greedy, *1900 Ref: https://codeforces.com/contest/1467/submission/103849987

+++ codeforces/1467/C.cpp
@@ -19,7 +19,7 @@ mt19937 rng(chrono::steady_clock::now().time_since_epoch().count());
 
 const int N = 1 << 19;
 int n[3];
-int a[3][N];
+ll a[3][N];
 ll sum[3];
 ll mn[3];
